content-script now sends the raw document to popup for it to parse, instead of sending the parsed content.
The response is serialised in transport, removing objects like Date instances.
Unfortunately this means we now parse the document twice, once in content script
to update the extension badge icon and again in the popup. Worth doing a shallow
parse in the content script, normal deep parse in the popup?