`<HCard/>` is now only expandable if it actually has additional content to show.
The collapsed `HCard` view shows a summary of a few core fields. Most `h-card`s are
sparse - just a name, maybe a link, maybe a photo... - so this summary is often
enough to show all the available data. In that case, th HCard no longer shows the
option to expand to its detailed view as there is no additional information to show.

Added `ExpandableCardProps.expandable`.